[PATCH 3] embedshim: rename utils to runtime_utils
And remove unuse codes.
